Update C++ API tests to use Catch2
